+++ app/local/substrate/lib/management.py
@@ -50,35 +50,53 @@ def fix_sys_path():
     """Fix the sys.path to include our extra paths."""
     sys.path = EXTRA_PATHS + SUBSTRATE_PATHS + USR_PATHS + sys.path
 
-def print_subcommand_overviews(commands):
+def print_subcommand_overviews(sub_comms, usr_comms):
     import logging
     logging.basicConfig(level=logging.ERROR)
-    print "manage.py commands: "
-    cmd_width = max(len(command) for command in commands)
-    for command in commands:
-        module = __import__("commands", {}, {}, [command])
+    print "manage.py built-in commands: "
+    cmd_width = max(len(command) for command in sub_comms)
+    for command in sub_comms:
+        module = __import__("substrate_manage.commands", {}, {}, [command])
         doc = getattr(module, command).__doc__
         print "  ", command.ljust(cmd_width), "-" if doc else "" ,  doc or ""
+    print "manage.py project commands: "
+    cmd_width = max(len(command) for command in usr_comms)
+    for command in usr_comms:
+        module = __import__("substrate_manage_usr.commands", {}, {}, [command])
+        doc = getattr(module, command).__doc__
+        print "  ", command.ljust(cmd_width), "-" if doc else "" ,  doc or ""
+
 
 def run_command(command, globals_, script_dir=SCRIPT_DIR):
     """Execute the file at the specified path with the passed-in globals."""
     fix_sys_path()
-    import pkgutil, commands
-    pkgpath = os.path.dirname(commands.__file__)
-    comms = [name for _, name, _ in pkgutil.iter_modules([pkgpath])]
+    import pkgutil
+    from substrate_manage import commands as substrate_commands
+    script_path = None
+    pkgpath = os.path.dirname(substrate_commands.__file__)
+    sub_comms = [name for _, name, _ in pkgutil.iter_modules([pkgpath])]
     for arg in sys.argv:
-        if arg in comms:
+        if arg in sub_comms:
+            script_path = './local/substrate/manage/substrate_manage/commands'
             break
     else:
-        print_subcommand_overviews(comms)
-        sys.exit(1)
+        from substrate_manage_usr import commands as usr_commands
+        pkgpath = os.path.dirname(usr_commands.__file__)
+        usr_comms = [name for _, name, _ in pkgutil.iter_modules([pkgpath])]
+        for arg in sys.argv:
+            if arg in usr_comms:
+                script_path = './local/usr/manage/substrate_manage_usr/commands'
+                break
+        else:
+            print_subcommand_overviews(sub_comms, usr_comms)
+            sys.exit(1)
     command_idx = sys.argv.index(arg)
     script_name = sys.argv[command_idx]
     management_args = sys.argv[:command_idx]
 
     command_args = sys.argv[command_idx:]
     sys.argv = command_args
-    script_path = os.path.join('./local/substrate/manage/commands', script_name + ".py")
+    script_path = os.path.join(script_path, script_name + ".py")
     execfile(script_path, globals_)